Move paging implementation to grid module
