+++ android/features/shows-grid/src/main/kotlin/com/thomaskioko/tvmaniac/show_grid/GridComposable.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.compose.components
+package com.thomaskioko.tvmaniac.show_grid
 
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Arrangement
@@ -6,12 +6,10 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.GridCells
-import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyGridScope
 import androidx.compose.foundation.lazy.LazyItemScope
 import androidx.compose.foundation.lazy.LazyListState
 import androidx.compose.foundation.lazy.LazyVerticalGrid
-import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.material.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -19,48 +17,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
-import kotlinx.coroutines.CoroutineScope
-
-@Composable
-fun <T> LazyGridItems(
-    items: List<T> = listOf(),
-    rows: Int = 3,
-    hPadding: Int = 8,
-    listState: LazyListState,
-    itemContent: @Composable LazyItemScope.(T) -> Unit
-) {
-    val chunkedList = items.chunked(rows)
-    LazyColumn(
-        state = listState,
-        modifier = Modifier
-            .padding(horizontal = hPadding.dp),
-        verticalArrangement = Arrangement.spacedBy(4.dp),
-    ) {
-
-        itemsIndexed(chunkedList) { _, item ->
-            Row(modifier = Modifier) {
-                item.forEachIndexed { _, item ->
-                    Box(
-                        modifier = Modifier
-                            .weight(1F)
-                            .align(Alignment.Top)
-                            .padding(2.dp),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        itemContent(item)
-                    }
-                }
-                repeat(rows - item.size) {
-                    Box(
-                        modifier = Modifier
-                            .weight(1F)
-                            .padding(8.dp)
-                    )
-                }
-            }
-        }
-    }
-}
+import com.thomaskioko.tvmaniac.compose.components.CircularLoadingView
+import com.thomaskioko.tvmaniac.compose.components.LoadingItem
+import com.thomaskioko.tvmaniac.compose.components.SnackBarErrorRetry
 
 @ExperimentalFoundationApi
 @Composable
@@ -68,7 +27,6 @@ fun <T : Any> LazyPagedGridItems(
     listState: LazyListState,
     lazyPagingItems: LazyPagingItems<T>,
     hostState: SnackbarHostState,
-    coroutineScope: CoroutineScope,
     rows: Int = 3,
     hPadding: Int = 2,
     itemContent: @Composable LazyItemScope.(value: T?) -> Unit
